# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sistema de Frequencia Avance",
    "version": "1.1.0",
    "description": "Documentacao oficial da API versionada em /api/v1 com envelope padrao de resposta."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versao atual"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success", "data", "message", "errors"],
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "nullable": true,
            "description": "Payload principal de sucesso."
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string", "nullable": true },
                "location": { "type": "string", "nullable": true },
                "message": { "type": "string" },
                "code": { "type": "string", "nullable": true },
                "value": { "nullable": true }
              }
            }
          }
        }
      },
...
      "SalaCreateRequest": {
        "type": "object",
        "required": ["nome", "turno"],
        "properties": {
          "nome": { "type": "string", "example": "Sala A" },
          "turno": { "type": "string", "example": "Manha" }
        }
      },
      "AlunoCreateRequest": {
        "type": "object",
        "required": ["nome"],
        "properties": {
          "nome": { "type": "string", "example": "Ana Clara" },
          "responsaveis": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["nome", "email", "telefone", "dataNascimento"],
              "properties": {
                "nome": { "type": "string" },
                "email": { "type": "string", "format": "email" },
                "telefone": { "type": "string" },
                "dataNascimento": { "type": "string", "format": "date" }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Api200": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "ok": {
                "value": {
                  "success": true,
                  "data": { "id": 1, "nome": "Joao" },
                  "message": "Operacao realizada com sucesso.",
                  "errors": null
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erro de validacao",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "validation": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Erro de validacao nos dados enviados.",
                  "errors": [
                    {
                      "field": "page",
                      "location": "query",
                      "message": "page deve ser um inteiro maior que 0.",
                      "value": "0"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Nao autenticado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "missing-token": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Token de acesso ausente.",
                  "errors": [
                    { "code": "AUTH_TOKEN_MISSING" }
                  ]
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/salas": {
      "get": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Lista salas com paginacao, filtros e ordenacao",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "turno", "in": "query", "schema": { "type": "string" } },
          { "name": "includeMeta", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["id", "nome", "turno", "totalAlunos"] } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "post": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Cria sala",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalaCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/salas/{salaId}": {
      "get": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Busca detalhes de uma sala",
        "parameters": [
          { "name": "salaId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "put": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Atualiza dados da sala",
        "parameters": [
          { "name": "salaId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalaCreateRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "delete": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Exclui sala",
        "parameters": [
          { "name": "salaId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "204": { "description": "Sem conteudo" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/salas/{salaId}/alunos": {
      "get": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Lista alunos de uma sala",
        "parameters": [
          { "name": "salaId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "includeMeta", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["id", "nome"] } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "post": {
        "tags": ["Salas"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Cadastra aluno na sala",
        "parameters": [
          { "name": "salaId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AlunoCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "409": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  }
}