# backend/routes/salas.routes.js
import { Router } from 'express';
import { body, query } from 'express-validator';
import pool from '../config/db.js';
import { validateRequest } from '../middlewares/validation.middleware.js';
import { broadcastRealtime } from '../realtime.js';
import {
  isSafeDisplayName,
  isValidDate,
  isValidEmail,
  isValidPhone,
  isValidTurno,
  normalizeEmail,
  normalizePhone,
  normalizeText,
  parsePagination,
  parsePositiveInt,
  sanitizeSearchTerm,
} from '../utils/validation.js';
import {
  paginationValidators,
  positiveIdParamValidator,
  sortByValidator,
} from '../validators/common.validators.js';

const router = Router();

const SORTABLE_SALAS_FIELDS = {
  id: 's.id',
  nome: 's.nome',
  turno: 's.turno',
  totalAlunos: 'total_alunos',
};

const SORTABLE_ALUNOS_FIELDS = {
  id: 'id',
  nome: 'nome',
};

function resolveSortField(sortableFields, sortBy, fallback) {
  return sortableFields[sortBy] || fallback;
}

function isMetaRequested(value) {
  return String(value || '').toLowerCase() === 'true';
}

function buildListMeta({ total, pagination, count }) {
  return {
    total,
    page: pagination?.page || 1,
    limit: pagination?.limit || count || 0,
    hasNextPage: pagination ? pagination.offset + count < total : false,
  };
}

router.get(
  '/',
  [
    ...paginationValidators,
    query('search').optional().isString().isLength({ max: 60 }).withMessage('search deve ter no maximo 60 caracteres.'),
    query('turno').optional().isString().isLength({ max: 30 }).withMessage('turno deve ter no maximo 30 caracteres.'),
    sortByValidator(Object.keys(SORTABLE_SALAS_FIELDS)),
    validateRequest,
  ],
  async (req, res) => {
  try {
    const search = sanitizeSearchTerm(req.query.search, 60);
    const turno = normalizeText(req.query.turno, 30);
    const includeMeta = isMetaRequested(req.query.includeMeta);
    const pagination = parsePagination(req.query);
    const sortBy = String(req.query.sortBy || 'nome');
    const sortOrder = String(req.query.order || 'asc').toUpperCase() === 'DESC' ? 'DESC' : 'ASC';
    const sortField = resolveSortField(SORTABLE_SALAS_FIELDS, sortBy, SORTABLE_SALAS_FIELDS.nome);
    const params = [];
    const where = [];

    if (search) {
      where.push('s.nome LIKE ?');
      params.push(`%${search}%`);
    }

    if (turno) {
      where.push('s.turno = ?');
      params.push(turno);
    }

    const whereSql = where.length > 0 ? `WHERE ${where.join(' AND ')}` : '';

    let total = null;
    if (includeMeta || pagination) {
      const [totalRows] = await pool.query(
        `SELECT COUNT(*) AS total
         FROM salas s
         ${whereSql}`,
        params,
      );
      total = Number(totalRows[0]?.total || 0);
    }

    const paginationSql = pagination ? 'LIMIT ? OFFSET ?' : '';
    const listParams = pagination ? [...params, pagination.limit, pagination.offset] : params;

    const [salas] = await pool.query(
      `SELECT s.id, s.nome, s.turno, COUNT(a.id) AS total_alunos
       FROM salas s
       LEFT JOIN alunos a ON a.sala_id = s.id
       ${whereSql}
       GROUP BY s.id
       ORDER BY ${sortField} ${sortOrder}
       ${paginationSql}`,
      listParams,
    );

    if (!includeMeta && !pagination) {
      return res.json(salas);
    }

    return res.json({
      items: salas,
      meta: buildListMeta({ total, pagination, count: salas.length }),
    });

  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao buscar salas.' });
  }
});

router.get('/:salaId', [positiveIdParamValidator('salaId', 'salaId'), validateRequest], async (req, res) => {
  try {
    const salaId = parsePositiveInt(req.params.salaId);

    if (!salaId) {
      return res.status(400).json({ message: 'Sala invalida.' });
    }

    const [rows] = await pool.query(
      `SELECT s.id, s.nome, s.turno, COUNT(a.id) AS total_alunos
       FROM salas s
       LEFT JOIN alunos a ON a.sala_id = s.id
       WHERE s.id = ?
       GROUP BY s.id
       LIMIT 1`,
      [salaId],
    );

    if (rows.length === 0) {
      return res.status(404).json({ message: 'Sala nao encontrada.' });
    }

    return res.json(rows[0]);
  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao buscar detalhes da sala.' });
  }
});

router.post(
  '/',
  [
    body('nome').trim().notEmpty().withMessage('nome e obrigatorio.'),
    body('turno').trim().notEmpty().withMessage('turno e obrigatorio.'),
    validateRequest,
  ],
  async (req, res) => {
  try {
    const nome = normalizeText(req.body?.nome, 80);
    const turno = normalizeText(req.body?.turno, 30);

    if (!nome || !turno) {
      return res.status(400).json({ message: 'Informe nome e turno da sala.' });
    }

    if (!isSafeDisplayName(nome, { min: 2, max: 80 }) || !isValidTurno(turno)) {
      return res.status(400).json({ message: 'Nome ou turno da sala contem caracteres invalidos.' });
    }

    const [result] = await pool.query(
      'INSERT INTO salas (nome, turno) VALUES (?, ?)',
      [nome, turno],
    );

    broadcastRealtime('salas.changed', {
      action: 'created',
      salaId: result.insertId,
      nome,
      turno,
    });

    return res.status(201).json({
      message: 'Sala criada com sucesso.',
      salaId: result.insertId,
    });
  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao criar sala.' });
  }
});

router.post(
  '/:salaId/alunos',
  [
    positiveIdParamValidator('salaId', 'salaId'),
    body('nome').trim().notEmpty().withMessage('nome do aluno e obrigatorio.'),
    body('responsaveis').optional().isArray().withMessage('responsaveis deve ser um array.'),
    validateRequest,
  ],
  async (req, res) => {
  let connection;
  let transactionStarted = false;

  try {
    connection = await pool.getConnection();

    const salaId = parsePositiveInt(req.params.salaId);
    const nome = normalizeText(req.body?.nome, 120);
    const responsaveis = Array.isArray(req.body?.responsaveis)
      ? req.body.responsaveis
      : [];

    if (!salaId) {
      return res.status(400).json({ message: 'Sala invalida.' });
    }

    if (!nome) {
      return res.status(400).json({ message: 'Informe o nome do aluno.' });
    }

    if (!isSafeDisplayName(nome, { min: 2, max: 120 })) {
      return res.status(400).json({ message: 'Nome do aluno contem caracteres invalidos.' });
    }

    const [salaRows] = await connection.query('SELECT id FROM salas WHERE id = ? LIMIT 1', [salaId]);
    if (salaRows.length === 0) {
      return res.status(404).json({ message: 'Sala nao encontrada.' });
    }

    const responsaveisNormalizados = [];
    const emailsNormalizados = new Set();

    for (const responsavel of responsaveis) {
      const nomeResponsavel = normalizeText(responsavel?.nome, 120);
      const email = normalizeEmail(responsavel?.email);
      const telefone = normalizePhone(responsavel?.telefone);
      const dataNascimento = normalizeText(responsavel?.dataNascimento, 10);

      if (!nomeResponsavel || !email || !telefone || !dataNascimento) {
        return res.status(400).json({ message: 'Cada responsavel precisa ter nome, email, telefone e data de nascimento.' });
      }

      if (!isSafeDisplayName(nomeResponsavel, { min: 2, max: 120 }) || !isValidEmail(email) || !isValidPhone(telefone) || !isValidDate(dataNascimento)) {
        return res.status(400).json({ message: 'Os dados do responsavel contem valores invalidos.' });
      }

      const emailKey = email.toLowerCase();
      if (emailsNormalizados.has(emailKey)) {
        return res.status(400).json({ message: 'Nao repita o mesmo email na lista de responsaveis.' });
      }

      emailsNormalizados.add(emailKey);
      responsaveisNormalizados.push({ nome: nomeResponsavel, email, telefone, dataNascimento });
    }

    await connection.beginTransaction();
    transactionStarted = true;

    const [result] = await connection.query(
      'INSERT INTO alunos (nome, sala_id) VALUES (?, ?)',
      [nome, salaId],
    );

    for (const responsavel of responsaveisNormalizados) {
      await connection.query(
        `INSERT INTO responsaveis (nome, email, telefone, data_nascimento, aluno_id)
         VALUES (?, ?, ?, ?, ?)`,
        [responsavel.nome, responsavel.email, responsavel.telefone, responsavel.dataNascimento, result.insertId],
      );
    }

    await connection.commit();

    broadcastRealtime('alunos.changed', {
      action: 'created',
      salaId,
      alunoId: result.insertId,
      totalResponsaveis: responsaveisNormalizados.length,
    });

    return res.status(201).json({
      message: responsaveisNormalizados.length > 0
        ? 'Aluno e responsaveis cadastrados com sucesso.'
        : 'Aluno cadastrado com sucesso.',
      alunoId: result.insertId,
      salaId,
      totalResponsaveis: responsaveisNormalizados.length,
    });
  } catch (error) {
    if (connection && transactionStarted) {
      await connection.rollback();
    }

    if (error?.code === 'ER_DUP_ENTRY') {
      return res.status(409).json({ message: 'Ja existe responsavel com este e-mail para este aluno.' });
    }

    console.error(error);
    return res.status(500).json({ message: 'Erro ao cadastrar aluno.' });
  } finally {
    if (connection) {
      connection.release();
    }
  }
});

router.get(
  '/:salaId/alunos',
  [
    positiveIdParamValidator('salaId', 'salaId'),
    ...paginationValidators,
    query('search').optional().isString().isLength({ max: 60 }).withMessage('search deve ter no maximo 60 caracteres.'),
    sortByValidator(Object.keys(SORTABLE_ALUNOS_FIELDS)),
    validateRequest,
  ],
  async (req, res) => {
  try {
    const salaId = parsePositiveInt(req.params.salaId);
    const search = sanitizeSearchTerm(req.query.search, 60);
    const includeMeta = isMetaRequested(req.query.includeMeta);
    const pagination = parsePagination(req.query);
    const sortBy = String(req.query.sortBy || 'nome');
    const sortOrder = String(req.query.order || 'asc').toUpperCase() === 'DESC' ? 'DESC' : 'ASC';
    const sortField = resolveSortField(SORTABLE_ALUNOS_FIELDS, sortBy, SORTABLE_ALUNOS_FIELDS.nome);

    if (!salaId) {
      return res.status(400).json({ message: 'Sala invalida.' });
    }

    const [salaRows] = await pool.query('SELECT id FROM salas WHERE id = ? LIMIT 1', [salaId]);
    if (salaRows.length === 0) {
      return res.status(404).json({ message: 'Sala nao encontrada.' });
    }

    const params = [salaId];
    const where = ['sala_id = ?'];

    if (search) {
      where.push('nome LIKE ?');
      params.push(`%${search}%`);
    }

    const whereSql = `WHERE ${where.join(' AND ')}`;

    let total = null;
    if (includeMeta || pagination) {
      const [totalRows] = await pool.query(
        `SELECT COUNT(*) AS total
         FROM alunos
         ${whereSql}`,
        params,
      );
      total = Number(totalRows[0]?.total || 0);
    }

    const paginationSql = pagination ? 'LIMIT ? OFFSET ?' : '';
    const listParams = pagination ? [...params, pagination.limit, pagination.offset] : params;

    const [alunos] = await pool.query(
      `SELECT id, nome
       FROM alunos
       ${whereSql}
       ORDER BY ${sortField} ${sortOrder}
       ${paginationSql}`,
      listParams,
    );

    if (!includeMeta && !pagination) {
      return res.json(alunos);
    }

    return res.json({
      items: alunos,
      meta: buildListMeta({ total, pagination, count: alunos.length }),
    });
  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao buscar alunos da sala.' });
  }
});

// Renomear sala
router.put(
  '/:salaId',
  [
    positiveIdParamValidator('salaId', 'salaId'),
    body('nome').trim().notEmpty().withMessage('nome e obrigatorio.'),
    body('turno').trim().notEmpty().withMessage('turno e obrigatorio.'),
    validateRequest,
  ],
  async (req, res) => {
  try {
    const salaId = parsePositiveInt(req.params.salaId);
    const nome = normalizeText(req.body?.nome, 80);
    const turno = normalizeText(req.body?.turno, 30);

    if (!salaId) {
      return res.status(400).json({ message: 'Sala inválida.' });
    }
    if (!nome || !turno) {
      return res.status(400).json({ message: 'Informe nome e turno da sala.' });
    }
    if (!isSafeDisplayName(nome, { min: 2, max: 80 }) || !isValidTurno(turno)) {
      return res.status(400).json({ message: 'Nome ou turno da sala contém caracteres inválidos.' });
    }

    const [result] = await pool.query(
      'UPDATE salas SET nome = ?, turno = ? WHERE id = ?',
      [nome, turno, salaId],
    );

    if (result.affectedRows === 0) {
      return res.status(404).json({ message: 'Sala não encontrada.' });
    }

    broadcastRealtime('salas.changed', {
      action: 'updated',
      salaId,
      nome,
      turno,
    });

    return res.json({ message: 'Sala atualizada com sucesso.' });
  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao atualizar sala.' });
  }
});

// Excluir sala
router.delete('/:salaId', [positiveIdParamValidator('salaId', 'salaId'), validateRequest], async (req, res) => {
  try {
    const salaId = parsePositiveInt(req.params.salaId);
    if (!salaId) {
      return res.status(400).json({ message: 'Sala inválida.' });
    }

    const [result] = await pool.query('DELETE FROM salas WHERE id = ?', [salaId]);
    if (result.affectedRows === 0) {
      return res.status(404).json({ message: 'Sala não encontrada.' });
    }

    broadcastRealtime('salas.changed', {
      action: 'deleted',
      salaId,
    });

    return res.status(204).send();
  } catch (error) {
    console.error(error);
    return res.status(500).json({ message: 'Erro ao excluir sala.' });
  }
});

export default router;